Fix: Dart errors, compileSdk 36, icon fixes
- Move dart:io import to top of camera_screen.dart
- Fix StencilPickerSheet import path
- Fix _ToolbarButton IconData vs Widget issue
- Change Icons.watermark to Icons.branding_watermark
- Fix extractPath type casting (num to double)
- Set compileSdk to 36

+++ lib/presentation/screens/editor/widgets/composite_preview.dart
@@ -232,8 +232,8 @@ class _StencilPreviewPainter extends CustomPainter {
       double distance = 0;
       while (distance < metric.length) {
         final start = distance;
-        final end = (distance + dashWidth).clamp(0, metric.length);
-        final extractPath = metric.extractPath(start, end);
+        final end = (distance + dashWidth).clamp(0.0, metric.length);
+        final extractPath = metric.extractPath(start, end.toDouble());
         canvas.drawPath(extractPath, paint);
         distance += dashWidth + dashSpace;
       }
